refactor(control): extract tool-result settlement into ToolEpisodeService's durable core
persistSettledToolResult, denyStandaloneTool, toolArgumentsExcerpt, and
observedSourceVersionsOf move from index.ts into
services/tool-episode-settlement.ts behind explicit ports (mutate, emit,
settleEffect). The exactly-once logical settlement now flows: parsed call
-> kernel/policy denial or execution -> EffectSettlementService.settle in
one place, with the composition root supplying only the mutex, event bus,
and effect boundary.

StandaloneToolSettlementInput becomes the owned contract of the
settlement module (type-aliased back into index.ts); the excerpt bound
constant moves with its excerpt function. No behavior change: same
artifacts, same tool states (PROPOSED/DENIED/SETTLED/FAILED), same
model-visible dual-path projection and truncation contract, same denial
JSON.

Characterization (tool-episode-settlement.test.ts): observed-source
cleaning bounds and filters sha256 observations and proves nothing on
failure; excerpts carry operands never contents and stay bounded; the
settlement input contract keeps artifact/capability/revision hooks
explicit.
